requirements.txt
anthropic>=0.52.0
claude-code-sdk>=0.1.0
pyyaml>=6.0
playwright>=1.40.0
